[#383] Refactor the export function, static, & lib to correct place
