+++ frontend/src/pages/cases/steps/SetIncomeTarget.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const SetIncomeTarget = () => {
+  return <div>SetIncomeTarget</div>;
+};
+
+export default SetIncomeTarget;
