+++ frontend/src/pages/cases/steps/index.js
@@ -1,0 +1,1 @@
+export { default as SetIncomeTarget } from "./SetIncomeTarget";
